requirements.txt
asgiref==3.3.1
bcrypt==3.2.0
cffi==1.14.5
cryptography==3.4.7
Django==3.1.7
Fabric3==1.14.post1
gunicorn==20.1.0
paramiko==2.7.2
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.16.0
pytz==2021.1
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
